move SelectOptionDetailed into components
