+++ components/SelectOptionDetailed.tsx
